statusの位置を変更
